move Configs from state module to config module
